Revert "Merge branch 'develop' into feature/misurare-tempo-progettazione-implementazione"
This reverts commit e76342c547cfd6004f0d6c680204fbb4ef72b52e, reversing
changes made to ed11ad7348585fcd7ee33df63dd19682f2efa4e1.
