--- database/dbConst.js
@@ -1,6 +1,0 @@
-module.exports = {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-    useCreateIndex: true,
-    useFindAndModify: false,
-};
