+++ database/models/category.js
@@ -1,9 +1,10 @@
 const mongoose = require('mongoose');
+const db = require('../dbConnect');
 
 const categorySchema = new mongoose.Schema({
   name: String,
   description: String
 })
 
-const Category = mongoose.model('Category', categorySchema);
+const Category = db.model('Category', categorySchema);
 module.exports = Category;
