+++ database/models/user.js
@@ -1,5 +1,6 @@
 const mongoose = require('mongoose');
 const JWT = require('jsonwebtoken');
+const db = require('../dbConnect');
 
 const UserSchema = new mongoose.Schema({
   name: String,
@@ -37,6 +38,6 @@ UserSchema.methods.generateToken = async function() {
   return token;
 }
 
-const User = mongoose.model('User', UserSchema);
+const User = db.model('User', UserSchema);
 
 module.exports = User;
