+++ index.js
@@ -1,5 +1,6 @@
-const app = require('./app');
-const mongoose = require('mongoose');
+const express = require("express");
+const app = express();
+const morgan = require('morgan');
 
 const usersRoute = require('./routes/users');
 const loginRoute = require('./routes/login');
@@ -16,16 +17,8 @@ app.use('/timers', timersRoute);
 app.use('/users', usersRoute);
 app.use('/login', loginRoute);
 app.use('/categories', categoriesRoute);
+
 const port = 3000;
-
-const dbOptions = require('./database/dbConst');
-
-mongoose.connect(process.env.DBURL, dbOptions).then(() => {
-  console.log("Connected to mongo");
-  app.listen(port, () => {
-    console.log(`Server listening on port ${port}`);
-  });
-})
 
 app.listen(port, () => {
   console.log(`Server listening on port ${port}`);
