--- utils/requestTypes.js
@@ -1,1 +1,0 @@
-module.exports.RequestTypes = ["illness","vacation"];
